Shapoorji Pallonji Vanaha Bavdhan Pune
Discover the true meaning of luxury living amidst greenery through the stature of not only the iconic Shapoorji Pallonji Vanaha Pune but also the location that puts you in the middle of urban conveniences. A world of luxury, intended to create an exclusive living experience, it is a work-of-art with luxury 2 & 3 BHK Flats that are a balance of space and style.

Shapoorji Vanaha Bavdhan Pune Highlight
Introduced by the world’s reputed developer Shapoorji Pallonji, #Vanaha is a residential community of high-rise apartments with the best possible use of utilization space and the best mixture of world-class amenities. Located in the prime area of Mumbai Highway, this community represents what Shapoorji Pallonji Real Estate is passionately crafting for people looking for an opulent standard of living. Every unit here is an integrated mix of space, ventilation, privacy, and security.

Being a part of 90 acres self-sustaining township, Shapoorji Vanaha Pune enjoys the best advantages. This grand residential community is spread over 6 acres of land, surrounded by an abundance of greenery, infrastructural and commercial developments, hotels & shops. A well-planned neighbourhood with 80% open space, green landscapes, and more.

In today’s busy life, it is becoming very hard to lead an active & social life after working hours & daily traffic jams. But not here, spend your quality time with relatives in the peaceful & stress-free environment of the clubhouse.

Positioned in the middle of refreshing greens and well-connected to Aundh, University Circle, Baner and Kothrud and every important location of Pune. Bavdhan is one of Pune’s most desirable localities & posh neighbourhoods make it a perfect address to own and reside in. Shapoorji Pallonji Vanaha is placed just 5  K.M. from the proposed Vanaz Metro Station on Line 2, as well as a host of happening landmarks of the city.

To know more - https://bit.ly/3EcI4vj